requirements.txt
selenium
webdriver-manager
fake-useragent
python-dotenv

